Add support for IPv6
The empty constructor for TcpSocket only supports IPv4, but getAddress can
return both IPv4 and IPv6 addresses (and maybe more). This defers instantiation
of std.socket.Socket until the address type is known and could be trivially
extended to try additional addresses if desired. This prevents TcpSocket.connect
from throwing an exception when an IPv6 address is passed into an IPv4
initialized TcpSocket.